backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
requests>=2.31
python-dotenv>=1.0
django-cors-headers>=4.0
